# src/routes/index.ts
import type { NextFunction, Request, Response } from 'express';
import express, { Router } from 'express';
import authRouter from './auth';
import userRouter from './user';
import deviceRouter from './device';
import hospitalRouter from './hospital';
import wardRouter from './ward';
import probeRouter from './probe';
import repairRouter from './repair';
import warrantyRouter from './warranty';
import notiRouter from './noti';
import logRouter from './log';
import configRouter from './config';
import firmwareRouter from './firmware';
import swaggerUi from 'swagger-ui-express';
import fs from 'node:fs';
import YAML from 'yaml';
import { backupData, getCompareDevice } from "../controllers";
import { BaseResponse } from '../models';
import { deviceEvent, historyList } from '../services';
import { verifyToken } from '../middlewares';

const file = fs.readFileSync("./swagger.yaml", "utf8");
const router = Router();

router.use('/auth', authRouter);
router.use('/user', userRouter);
router.use('/device', deviceRouter);
router.use('/config', configRouter);
router.use('/hospital', hospitalRouter);
router.use('/ward', wardRouter);
router.use('/probe', probeRouter);
router.use('/repair', repairRouter);
router.use('/warranty', warrantyRouter);
router.use('/notification', notiRouter);
router.use('/log', logRouter);
router.use('/firmwares', firmwareRouter);
router.use('/img', express.static('public/images'));
router.use('/font', express.static('public/fonts'));
router.use('/firmware', express.static('public/firmwares'));
router.use('/api-docs', swaggerUi.serve, swaggerUi.setup(YAML.parse(file)));
router.use('/backup', backupData);
router.use('/compare', verifyToken, getCompareDevice);
router.use('/history', verifyToken, async (req: Request, res: Response<BaseResponse>, next: NextFunction) => {
  try {
    res.status(200).json({ 
      message: 'Successful',
      success: true, 
      data: await historyList(res.locals.token)
    });
  } catch (error) {
    next(error);
  }
});
router.post('/mqtt/event', async (req: Request, res: Response<BaseResponse<Promise<string>>>, next: NextFunction) => {
  try {
    type EventDevice = {
      clientid: string,
      event: string
    };
    const data = req.body as EventDevice;
    res.status(200).json({ 
      message: 'Successful',
      success: true, 
      data: deviceEvent(data.clientid, data.event)
    });
  } catch (error) {
    next(error);
  }
});
router.use('/', (req: Request, res: Response<BaseResponse<null>>) => {
  res.status(404).json({ 
    message: 'Not Found',
    success: false, 
    data: null
  });
});

export default router;