# swagger.yaml
openapi: 3.0.0
info:
  title: E-TEMP API
  description: testing api page.
  version: 0.1.9
servers:
  - url: https://api.siamatic.co.th/etemp
  - url: http://localhost:8080/etemp
security:
  - bearerAuth: []
...
paths:
...
  /device:
    get:
      tags:
        - Device
      summary: Get a list of devices
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
    post:
      tags:
        - Device
      summary: Create a new device
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                fileupload:
                  type: string
                  format: binary
                body:
                  $ref: '#/components/schemas/Device'
      responses:
        '201':
          description: Device created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Device'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  # optional, for documentation purposes only
  schemas:
    Device:
      type: object
      properties:
        devId:
          type: string
        devName:
          type: string
        devSerial:
          type: string
        devStatus:
          type: string
        devType:
          type: string
        devModel:
          type: string
        devManufacturer:
          type: string
        devFirmware:
          type: string
        devHardware:
          type: string
        devLocation:
          type: string
        devDescription:
          type: string
        devImage:
          type: string
        devCreatedAt:
          type: string
          format: date-time
        devUpdatedAt:
          type: string
          format: date-time                    